Release 1.1
* Added: Internal Lagger errors/exceptions support
* Added: Support of Lagger_ActionException
* Added: Lagger_ActionSms by SMS gateway class of service http://litesms.net
* Added: Using custom functions as modifiers for templates(without registration)
* Improved: Defining templates like a strings, not like objects
* Improved: Lagger_ActionFileLog limits checking only in one of 100 iterations
* Fixed: Lagger_PrintAction now work correctly on flushing buffer
* Fixed: Correct tags matching check in Lagger_HandlerDebug
* Fixed: Lagger_Skiper expire time updating when expire check

+++ trunk/library/Lagger/ActionFileLog.php
@@ -1,35 +1,45 @@
 <?php
 
-class Lagger_ActionFileLog extends Lagger_ActionAbstract {
-
+/**
+ * 
+ * @see http://code.google.com/p/lagger
+ * @author Barbushin Sergey http://www.linkedin.com/in/barbushin
+ * 
+ */
+class Lagger_ActionFileLog extends Lagger_Action {
+	
 	protected $filepath;
 	protected $template;
 	protected $sizeLimit;
 	protected $daysLimit;
+	
+	const checkLimit = 100;
 
-	public function __construct(Lagger_Template $template, $filepath, $sizeLimit = null, $daysLimit = null) {
-		$this->filepath = $filepath;
+	public function __construct($template, $filepath, $sizeLimit = null, $daysLimit = null) {
+		$this->filepath = realpath($filepath);
 		$this->template = $template;
 		$this->sizeLimit = (int)$sizeLimit;
 		$this->daysLimit = (int)$daysLimit;
 	}
 
 	protected function make() {
-		$fp = fopen($this->filepath, 'a'); // TODO: lock check/set
+		$this->checkLimits();
 		
+		$fp = fopen($this->filepath, 'a'); // TODO: lock check/set
+		fputs($fp, $this->eventspace->fetch($this->template));
+		fclose($fp);
+	}
 
-		if ($this->sizeLimit || $this->daysLimit) {
-			$fstat = fstat($fp);
-			
-			// if(($this->daysLimit && (time() - $fstat['mtime']) > ($this->daysLimit * 24 * 60 * 60)) || ($this->sizeLimit && $fstat['size'] > $this->sizeLimit)) // TODO: check mdtime against cached ctime.. need some other
-			if ($this->daysLimit && (time() - $fstat['mtime']) > ($this->daysLimit * 24 * 60 * 60)) {
+	protected function checkLimits() {
+		if (!mt_rand(0, self::checkLimit)) {
+			if ($this->sizeLimit || $this->daysLimit) {
+				$fp = fopen($this->filepath, 'r');
+				$fstat = fstat($fp);
+				if ($this->daysLimit && (time() - $fstat['mtime']) > ($this->daysLimit * 24 * 60 * 60)) {
+					unlink($this->filepath); // TODO: lock check/set, no unlink - just cut to 1/2 ?
+				}
 				fclose($fp);
-				unlink($this->filepath); // TODO: lock check/set, no unlink - just cut to 1/2 ?
-				$fp = fopen($this->filepath, 'a');
 			}
 		}
-		
-		fputs($fp, $this->template->compile());
-		fclose($fp);
 	}
 }
